backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pymongo==4.6.0
python-dotenv==1.0.0
openai==0.28.0
anthropic==0.5.0
google-generativeai==0.3.0
